requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0  
Django==4.2.2
django-extensions==3.2.3   
djangorestframework==3.14.0
idna==3.4
joblib==1.3.2
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2     
pytz==2023.3
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0        
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.0.3
xgboost==1.7.6